Create README.md
I will introduce a home industry skill in Japan, namely, how to make “Washi is sometimes called ‘rice paper,’ even though it has nothing to do with rice.” A missionary in 1976 in an article taken from a Japanese correspondent, he said: "The work of making Washi begins with a paper mulberry tree," the fibers are taken from the inner bark, then thoroughly cleaned. Then the slimy liquid from the hibiscus root is added. The long fibers of the paper mulberry tree and the fluid from the hibiscus distinguish washi from other papers. The slimy substance ensures an even distribution of the fibers. It is a thrill to see this product. The long fibers of the paper mulberry tree and the fluid from the hibiscus tree distinguish washi from other papers.